% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatAll.R
\name{concatAll}
\alias{concatAll}
\title{Concact multiple values}
\usage{
concatAll(..., sep = "; ")
}
\arguments{
\item{...}{the list of values to be concacted}

\item{sep}{the separated for concating}
}
\description{
Concact multiple values
}
\examples{

a = 5
b = 10
c = "done"
concatAll(a,b,c)


}
