% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Define words in english}
\usage{
define(term, limitResult = 5)
}
\arguments{
\item{term}{The term for which the definition is requested}

\item{limitResult}{The limit of how many results to show}
}
\value{
Definition of word if it exisits
}
\description{
Given a word, a mini-dictionary is searched and results are retrieved with the relevant definitions
}
\examples{
\dontrun{
wordtodefine <- "smile"
define(wordtodefine)
}

}
