% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineSelected.R
\name{defineSel}
\alias{defineSel}
\title{Define Selected Text}
\usage{
defineSel()
}
\value{
definition of the selected text
}
\description{
Detects a selected word in an open document. This was written for Rmarkdown writing, but it can be used for an open document.
Using the addin is the most preferred means to call this function, albeit it can be called directly from the console.
}
\examples{
\dontrun{
defineSel()
}
}
