% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSelectedText.R
\name{fetchSelectedText}
\alias{fetchSelectedText}
\title{Fetch Selected Text}
\usage{
fetchSelectedText()
}
\value{
The selected text and if it contains alphabets
}
\description{
Fetch Selected Text
}
\examples{
\dontrun{
fetchSelectedText()
}

}
