% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDictionaryLoaded.R
\name{isDictionaryLoaded}
\alias{isDictionaryLoaded}
\title{Check if dictionary is loaded}
\usage{
isDictionaryLoaded()
}
\value{
Boolean result
}
\description{
Check if dictionary is loaded
}
\examples{
\dontrun{
y = isDictionaryLoaded()
y
}

}
