% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDictionary.R
\name{loadDictionary}
\alias{loadDictionary}
\title{Load the dictionary into the global environment}
\usage{
loadDictionary()
}
\description{
Load the dictionary into the global environment
}
\examples{

loadDictionary()

}
