% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{splitSpa12}
\alias{splitSpa12}
\title{Split definition character and count}
\usage{
splitSpa12(word)
}
\arguments{
\item{word}{the fullword to split}
}
\value{
whole word from the given text
}
\description{
Split definition character and count
}
\examples{
\dontrun{
wordtodefine <- "smile v"
splitSpa12(wordtodefine)
}

}
